backend/requirements.txt
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Django==5.0.7
psycopg2-binary==2.9.9
sqlparse==0.5.1
djangorestframework==3.15.2
pillow==10.4.0
djangorestframework-simplejwt==5.3.1
redis==5.1.1
pytz==2024.2
django-redis==5.4.0
